[issue-456] move actor_parser.py as it is format-agnostic and used for json,yaml,xml and rdf parsing
